fix: Fix status library headers and move test
