+++ packager/status/status_test_util.h
@@ -9,7 +9,7 @@
 
 #include <gtest/gtest.h>
 
-#include "packager/status.h"
+#include "packager/status/status.h"
 
 #define EXPECT_OK(val) EXPECT_EQ(shaka::Status::OK, (val))
 #define ASSERT_OK(val) ASSERT_EQ(shaka::Status::OK, (val))
