+++ packager/status/status_test_util_unittest.cc
@@ -6,7 +6,7 @@
 
 #include <gtest/gtest-spi.h>
 
-#include "packager/status_test_util.h"
+#include "packager/status/status_test_util.h"
 
 namespace shaka {
 namespace media {
